Added more background options, added relative action support.
Does not compile due to new language system.

git-svn-id: https://svn.code.sf.net/p/gjava/code@1490 0bf6c8e7-452b-0410-8c87-9aaedaf074f6

+++ Aurora/src/org/gcreator/units/BackgroundInScene.java
@@ -25,7 +25,7 @@ public class BackgroundInScene implements Serializable{
     public int hmode = MODE_SINGLE;
     public int vmode = MODE_SINGLE;
     public boolean visibleonstart = false,stretch=false;
-    public int xpos=0,ypos=0,hspeed=0,vspeed=0,tileh=0,tilev=0;
+    public double xpos=0,ypos=0,hspeed=0,vspeed=0,tileh=0,tilev=0;
     
     public BackgroundInScene(String name){
         this.name = name;
